+++ docs/versionhistory.rst
@@ -33,8 +33,7 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Added the ``anyio.Path.is_junction()`` and ``anyio.Path.walk()`` methods
 - Added ``create_unix_datagram_socket`` and ``create_connected_unix_datagram_socket`` to
   create UNIX datagram sockets (PR by Jean Hominal)
-- Added the ``deadline_reached`` property to ``CancelScope`` to determine whether the
-  scope was cancelled due to the deadline being reached, or for some other reason
+- Added the ``cancelled_caught`` property to ``CancelScope``
 - Fixed ``fail_after()`` raising an unwarranted ``TimeoutError`` when the cancel scope
   was cancelled before reaching its deadline